support 32 uint
No significant difference in performance

+++ lib/type_conversion/base_types.cpp
@@ -112,10 +112,10 @@ void BaseTypes::injectIntegerNumber(const std::string & value, uint8_t n_bytes, 
     _injectUnsignedNumber(number);
 }
 
-int32_t BaseTypes::extractIntegerNumber(uint8_t n_bytes, bool is_unsigned) {
-    uint32_t output_data = 0;
-    int32_t int_number = 0;
-    int32_t sign_factor = 1;
+int64_t BaseTypes::extractIntegerNumber(uint8_t n_bytes, bool is_unsigned) {
+    uint64_t output_data = 0;
+    int64_t int_number = 0;
+    int64_t sign_factor = 1;
     if (!is_unsigned) {
         output_data = bit_stream_->get_max_4bytes(1);
         if (output_data) {
